+++ app/components/oversee/essence.rb
@@ -1,6 +1,4 @@
 class Oversee::Essence < Phlex::HTML
-  extend Phlex::Kit
-
   TAILWIND_MERGER = ::TailwindMerge::Merger.new.freeze unless defined?(TAILWIND_MERGER)
 
   attr_reader :attributes