+++ lib/oversee.rb
@@ -8,6 +8,10 @@ require "oversee/configuration"
 
 # Phlex
 require "phlex-rails"
+# module Views
+# end
+# module ::Components
+# end
 
 # Pagy
 require "pagy"
@@ -44,14 +48,5 @@ module Oversee
       files = Dir.glob(root.join("**/*.rb"))
       files.map! { |f| f.split(root.to_s).last.delete_suffix(".rb").classify.prepend("Cards::") }
     end
-  end
-
-  # VIEWS
-  module Views
-  end
-
-  # COMPONENTS
-  module Components
-    extend Phlex::Kit
   end
 end