+++ CHANGELOG.md
@@ -4,10 +4,12 @@ This is the changelog for the `oversee` gem. For now, most of the updates will b
 
 ### Unreleased
 
+- Introducing `Oversee::Entity` and `Oversee::Record` classes
 - Componentizing associated resources
 - Adding `essence` component library to make the UI streamlined
 - Making boolean input friendlier - using toggle switch
 - Fixing incorrect linking for has_many associations
+- Removing away `Phlex::Kit`. We'll write it out by hand.
 
 ### 0.3.1